+++ R/combine_fude.R
@@ -5,26 +5,24 @@
 #' Fude Polygon data to the community units.
 #'
 #' @param data
-#'   List of MAFF Fude Polygon data.
+#'   Fude Polygon data as returned by [read_fude()].
 #' @param boundary
-#'   List of one or more MAFF agricultural community boundary data.
+#'   Agricultural community boundary data as returned by [get_boundary()].
 #' @param city
-#'   A local government name in Japanese to be extracted. In the case of
-#'   overlapping local government names, this must contain the prefecture name
-#'   in Japanese and the prefecture code in romaji (e.g., "Fuchu-shi, 13",
-#'   "fuchu 13",  "34 fuchu-shi",  "34, FUCHU-CHO"). Alternatively, it could be
-#'   a 6-digit local government code.
+#'   A character vector of local government names or 6-digit local government
+#'   codes to extract.
 #' @param kcity
-#'   A regular expression string. One or more former city names (in Japanese)
+#'   A regular expression. One or more former municipality names (in Japanese)
 #'   to extract.
 #' @param rcom
-#'   A regular expression string. One or more agricultural community names (in
+#'   A regular expression. One or more agricultural community names (in
 #'   Japanese) to extract.
 #' @param year
-#'   Year in the column name of the `data`. If there is more than one
-#'   applicable local government code, it is required.
+#'   Year in the column name of the `data`. If there is more than one applicable
+#'   local government code, it is required.
 #'
-#' @returns A list of [sf::sf()] objects.
+#' @returns
+#'   A list of [sf::sf()] objects.
 #'
 #'  @seealso [read_fude()].
 #'