+++ R/extract_boundary.R
@@ -1,27 +1,26 @@
 #' Extract specified agricultural community boundary data
 #'
 #' @description
-#' `extract_boundary()` extracts the specified data from the list returned by
-#' [get_boundary()].
+#' `extract_boundary()` extracts specified subsets of agricultural community
+#' boundary data returned by [get_boundary()].
 #'
 #' @param boundary
-#'   List of one or more MAFF agricultural community boundary data.
+#'   Agricultural community boundary data as returned by [get_boundary()].
 #' @param city
-#'   A local government name in Japanese to be extracted. In the case of
-#'   overlapping local government names, this must contain the prefecture name
-#'   in Japanese and the prefecture code in romaji (e.g., "Fuchu-shi, 13",
-#'   "fuchu 13",  "34 fuchu-shi",  "34, FUCHU-CHO"). Alternatively, it could be
-#'   a 6-digit local government code.
+#'   A character vector of local government names or 6-digit local government
+#'   codes to extract.
 #' @param kcity
-#'   A regular expression string. One or more former city names (in Japanese)
+#'   A regular expression. One or more former municipality names (in Japanese)
 #'   to extract.
 #' @param rcom
-#'   A regular expression string. One or more agricultural community names (in
+#'   A regular expression. One or more agricultural community names (in
 #'   Japanese) to extract.
 #' @param layer
-#'   Logical.
+#'   Logical. If `TRUE`, the returned object includes not only agricultural
+#'   community boundaries but also prefecture and municipality boundaries.
 #'
-#' @returns A list of [sf::sf()] object(s).
+#' @returns
+#'   An [sf::sf()] object.
 #'
 #' @seealso [read_fude()].
 #'